Bug 1621731 - Improve Mercurial submit performance on large repos; r=zalun
Instead of generating a list of file mode changes for all files in the
repository, which is expensive on mozilla-central's ~300k files, use
the list of files modified by the commit being submitted.

On my system this reduces the time `_get_file_modes` takes on a
mozilla-central repo from 17.5 seconds down to 1.5 seconds.

Differential Revision: https://phabricator.services.mozilla.com/D67803